+++ Modules/Microsoft365DSC/Modules/M365DSCUtil.psm1
@@ -1293,7 +1293,51 @@ function Import-M365DSCDependencies
 
     foreach ($dependency in $dependencies)
     {
-        Import-Module $dependency.ModuleName -Force
+        Import-Module $dependency.ModuleName -RequiredVersion $dependency.RequiredVersion -Force
+    }
+}
+
+<#
+.Description
+This function removes all versions of dependencies that are not specified in the manifest from the current PowerShell session.
+
+.Example
+Remove-M365DSCInvalidDependenciesFromSession
+
+.Functionality
+Private
+#>
+function Remove-M365DSCInvalidDependenciesFromSession
+{
+    [CmdletBinding()]
+    param()
+
+    $currentPath = Join-Path -Path $PSScriptRoot -ChildPath '..\' -Resolve
+    $manifest = Import-PowerShellDataFile "$currentPath/Dependencies/Manifest.psd1"
+    $dependencies = $manifest.Dependencies
+
+    foreach ($dependency in $dependencies)
+    {
+        $loadedModuleInstances = Get-Module $dependency.ModuleName
+
+        $incorrectModuleVersions = $null
+        if ($loadedModuleInstances)
+        {
+            $incorrectModuleVersions = $loadedModuleInstances | Where-Object -FilterScript { $_.Version -ne $dependency.RequiredVersion }
+
+            if ($incorrectModuleVersions)
+            {
+                foreach ($incorrectVersion in $incorrectModuleVersions)
+                {
+                    $FQN = @{
+                        ModuleName    = $incorrectVersion.Name
+                        ModuleVersion = $incorrectVersion.Version
+                    }
+                    Write-Verbose -Message "Removing Module {$($incorrectVersion.Name)} version {$($incorrectVersion.Version)} from the current PowerShell session"
+                    Remove-Module -FullyQualifiedName $FQN -Force -ErrorAction SilentlyContinue
+                }
+            }
+        }
     }
 }
 
@@ -2363,6 +2407,13 @@ function Assert-M365DSCBlueprint
                 $ResourcesInBluePrint += $resource.ResourceName
             }
         }
+
+        if (!$ResourcesInBluePrint)
+        {
+            Write-Host "Malformed BluePrint, aborting"
+            break
+        }
+
         Write-Host "Selected BluePrint contains ($($ResourcesInBluePrint.Length)) components to assess."
 
         # Call the Export-M365DSCConfiguration cmdlet to extract only the resource
@@ -2951,6 +3002,13 @@ function Test-M365DSCNewVersionAvailable
 {
     [CmdletBinding()]
     param()
+
+    if ("AzureAutomation/" -eq $env:AZUREPS_HOST_ENVIRONMENT)
+    {
+        $message = "Skipping check for newer version of Microsoft 365 DSC due to Azure Automation Environment restrictions."
+        Write-Verbose -Message $message
+        return
+    }
 
     try
     {
@@ -3640,9 +3698,6 @@ function New-M365DSCMissingResourcesExample
 .Description
 This function validates there are no updates to the module or it's dependencies and no multiple versions are present on the local system.
 
-.Parameter Force
-Specifies that all dependencies should be forcefully imported again.
-
 .Example
 Test-M365DSCModuleValidity
 
@@ -3657,6 +3712,14 @@ function Test-M365DSCModuleValidity
     [CmdletBinding()]
     param(
     )
+
+    if ("AzureAutomation/" -eq $env:AZUREPS_HOST_ENVIRONMENT)
+    {
+        $message = "Skipping check for newer version of Microsoft 365 DSC due to Azure Automation Environment restrictions."
+        Write-Verbose -Message $message
+        return
+    }
+
     $InformationPreference = 'Continue'
 
     # validate only one installation of the module is present (and it's the latest version available from the psgallery)